refactor: rename Sorting to Filters
